+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/agent/MyStudioScreen.kt
@@ -473,40 +473,26 @@ fun ToolMakerListView(
                         toolMaker,
                         onDismissRequest = { showEditServerConfigDialog=false },
                         onConfirmRequest = {
-//                            myStudioViewModel.modifyMCPServerConfigForStudio(
-//                                toolAgent,toolMaker,it
-//                            ){ code, message, mcpServer ->
-//                                if(code==0) {
-//                                    val config = AIPortMCPServerConfig()
-//                                    config.id = toolMaker.id
-//                                    config.transport = it.transport
-//                                    config.url = it.url
-//                                    config.command = it.command
-//                                    config.args = JSON.encodeToJsonElement(it.args).toString()
-//                                    config.env = JSON.encodeToJsonElement(it.env).toString()
-//                                    myStudioViewModel.modifyMCPServerConfig(
-//                                        myStudioViewModel.toolAgent,
-//                                        config
-//                                    )
-//                                }else{
-//                                    generalViewModel.showSnackbar(
-//                                        message?:"Config MCP Server ${toolMaker.name} Error",
-//                                        actionLabel = "Error",
-//                                        withDismissAction = true
-//                                    )
-//                                }
-//                            }
-                            val config = AIPortMCPServerConfig()
-                            config.id = toolMaker.id
-                            config.transport = it.transport
-                            config.url = it.url
-                            config.command = it.command
-                            config.args = JSON.encodeToJsonElement(it.args).toString()
-                            config.env = JSON.encodeToJsonElement(it.env).toString()
-                            myStudioViewModel.modifyMCPServerConfig(
-                                myStudioViewModel.toolAgent,
-                                config
-                            )
+                            mcpServer,mcpServerCOnfig->
+                            if(mcpServer.id>Int.MAX_VALUE) {
+                                val config = AIPortMCPServerConfig()
+                                config.id = toolMaker.id
+                                config.transport = mcpServerCOnfig.transport
+                                config.url = mcpServerCOnfig.url
+                                config.command = mcpServerCOnfig.command
+                                config.args = JSON.encodeToJsonElement(mcpServerCOnfig.args).toString()
+                                config.env = JSON.encodeToJsonElement(mcpServerCOnfig.env).toString()
+                                myStudioViewModel.modifyMCPServerConfig(
+                                    myStudioViewModel.toolAgent,
+                                    config
+                                )
+                            }else{
+                                myStudioViewModel.modifyMCPServerConfigForStudio(
+                                    myStudioViewModel.toolAgent,
+                                    mcpServer,
+                                    mcpServerCOnfig
+                                )
+                            }
                         }
                     )
                 }